fs: Move out header into its own file.
